+++ linen-repository/src/main/scala/x7c1/linen/repository/dummy/TraceableQueue.scala
@@ -1,7 +1,7 @@
 package x7c1.linen.repository.dummy
 
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.repository.crawler.{InspectedSource, SourceDequeueEvent, SourceLoader, SourceUpdaterQueue, UpdatedSource}
+import x7c1.linen.repository.loader.crawling.{InspectedSource, SourceLoader, UpdatedSource, SourceDequeueEvent, SourceUpdaterQueue}
 import x7c1.wheat.macros.logger.Log
 
 import scala.collection.mutable
