+++ linen-repository/src/main/scala/x7c1/linen/repository/loader/crawling/InvalidEntry.scala
@@ -1,4 +1,4 @@
-package x7c1.linen.repository.crawler
+package x7c1.linen.repository.loader.crawling
 
 sealed trait InvalidEntry
