+++ linen-repository/src/main/scala/x7c1/linen/repository/loader/crawling/SourceUpdaterQueue.scala
@@ -1,4 +1,4 @@
-package x7c1.linen.repository.crawler
+package x7c1.linen.repository.loader.crawling
 
 import java.lang.System.currentTimeMillis
