+++ linen-repository/src/test/scala/x7c1/linen/repository/loader/crawling/SourceUpdaterQueueTest.scala
@@ -1,4 +1,4 @@
-package x7c1.linen.repository.crawler
+package x7c1.linen.repository.loader.crawling
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
