+++ linen-scene/src/main/scala/x7c1/linen/scene/updater/UpdaterMethods.scala
@@ -4,9 +4,9 @@ import android.app.Service
 import android.content.Intent
 import x7c1.linen.database.control.DatabaseHelper
 import x7c1.linen.glue.service.ServiceControl
-import x7c1.linen.repository.crawler.{Implicits, SourceInspector}
 import x7c1.linen.repository.date.Date
 import x7c1.linen.repository.dummy.{DummyFactory, TraceableQueue}
+import x7c1.linen.repository.loader.crawling.{Implicits, SourceInspector}
 import x7c1.linen.repository.preset.PresetFactory
 import x7c1.linen.repository.source.setting.{SettingSource, SettingSourceAccessorFactory}
 import x7c1.wheat.macros.intent.{ExtraNotFound, IntentExpander}
